Extract the StatusManager into its own public class
...and add a hideAll() method.
